+++ kwcoco/__init__.py
@@ -269,7 +269,7 @@ Testing:
 
 """
 
-__version__ = '0.5.9'
+__version__ = '0.6.0'
 
 
 __submodules__ = {